requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
jinja2==3.1.3
python-jose==3.3.0
python-dotenv==1.0.1
supabase==1.2.0
httpx==0.24.1
gotrue==1.3.1
bcrypt==4.1.2
sniffio==1.3.1
anyio==4.2.0
starlette==0.36.3
typing-extensions==4.9.0
pydantic==2.6.1
pydantic-core==2.16.2
email-validator==2.1.0.post1
idna==3.6
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
click==8.1.7
colorama==0.4.6
watchfiles==0.21.0
websockets==12.0
aiosmtplib==3.0.1